+++ packages/config/src/constants/index.ts
@@ -1,3 +1,5 @@
 export * from './ipfs'
 export * from './paths'
-export * from './tokens'
+export * from './addresses'
+export * from './nativeCurrencyTemplate'
+export * from './wrappedTokens'